+++ src/HrHub.Domain/Entities/SqlDbEntities/ContentComment.cs
@@ -2,12 +2,12 @@
 
 namespace HrHub.Domain.Entities.SqlDbEntities
 {
-    public class ContentComment : TypeCardEntity<int>
+    public class ContentComment : TypeCardEntity<long>
     {
-        public int ContentId { get; set; }
+        public long ContentId { get; set; }
         public bool Pinned { get; set; }
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public DateTime ContentDate { get; set; }
-        public int MasterContentId { get; set; }
+        public long MasterContentId { get; set; }
     }
 }
